chore(release): harden storage adapter

+++ test/test_helper.exs
@@ -1,10 +1,24 @@
 ExUnit.start()
 
-db_path = Application.fetch_env!(:jido_ecto, Jido.Ecto.TestRepo)[:database]
-File.mkdir_p!(Path.dirname(db_path))
-File.rm(db_path)
+repo_config = Application.fetch_env!(:jido_ecto, Jido.Ecto.TestRepo)
+
+if repo_config[:adapter] == Ecto.Adapters.SQLite3 do
+  db_path = Keyword.fetch!(repo_config, :database)
+  File.mkdir_p!(Path.dirname(db_path))
+  File.rm(db_path)
+end
 
 {:ok, _pid} = Jido.Ecto.TestRepo.start_link()
+
+if repo_config[:adapter] == Ecto.Adapters.Postgres do
+  for statement <- [
+        "DROP TABLE IF EXISTS jido_thread_entries",
+        "DROP TABLE IF EXISTS jido_threads",
+        "DROP TABLE IF EXISTS jido_checkpoints"
+      ] do
+    Ecto.Adapters.SQL.query!(Jido.Ecto.TestRepo, statement, [])
+  end
+end
 
 {:ok, _migrated, _apps} =
   Ecto.Migrator.with_repo(Jido.Ecto.TestRepo, fn repo ->
